+++ src/hooks/useSlideNav.tsx
@@ -1,4 +1,3 @@
-import { db } from 'firebase/clientApp'
 import { useCallback, useRef, useState } from 'react'
 
 type ChatTable = {
@@ -6,20 +5,10 @@ type ChatTable = {
   id: string
 }[]
 
-type FormProps = {
-  theme: string
-  name: string
-  date: string
-  time: string
-  link: string
-  discription: string
-}
-
 export const useSlideNav = () => {
   const [chatText, setChatText] = useState('')
   const [isChatOpen, isNotChatOpen] = useState(false)
   const [chatTable, setChatTable] = useState<ChatTable>([])
-  const [data, setData] = useState({ theme: '', name: '', date: '', time: '', link: '', discription: '' })
 
   const scrollRef = useRef<HTMLDivElement>(null)
 
@@ -27,26 +16,6 @@ export const useSlideNav = () => {
     scrollRef?.current?.scrollIntoView()
     console.log('ugoita')
   }, [scrollRef])
-
-  const hundleAdd = async (data: FormProps) => {
-    await db.collection('contents').doc(data.theme).set({
-      theme: data.theme,
-      name: data.name,
-      date: data.date,
-      time: data.time,
-      link: data.link,
-      discription: data.discription,
-    })
-    setData({ theme: '', name: '', date: '', time: '', link: '', discription: '' })
-  }
-
-  const hundleChange = useCallback(
-    (e) => {
-      const value = e.target.value
-      setData({ ...data, [e.target.name]: value })
-    },
-    [data]
-  )
 
   const hundleText = useCallback(
     (e) => {
@@ -67,9 +36,6 @@ export const useSlideNav = () => {
     hundleChat,
     chatTable,
     setChatTable,
-    data,
-    hundleChange,
-    hundleAdd,
     scrollRef,
     scrollChatList,
   }
